Current Logisoso, as UFO
Based on the archive downloadable on the foundry.
I took the liberty of removing the ttf and the vfb,
and generating a UFO from the vfb.

I don’t think we need to version both the vfb and the UFO
at the same time (except in cases where the UFO format
proves to be lacking in relation to the vfb).